MIN_VALUE has different semantics for floating point types than integer types
Float.MIN_VALUE is the smallest positive Float value larger than 0, not the smallest Float value. 

The splitter has a further problem in that in the degenerate case all objects in a leaf with the same bounding boxes it will crash because seed1 and seed2 is null, these should simply be initialized to any two objects from the leaf.